# Measure compile time only when using time macros pt.2: handling when code under test throws and make compile timing thread-local