--- a/base/timing.jl
+++ b/base/timing.jl
@@ -124,45 +124,42 @@ function format_bytes(bytes) # also used by InteractiveUtils
     end
 end
 
-const print_lock = ReentrantLock()
 function time_print(elapsedtime, bytes=0, gctime=0, allocs=0, compile_time=0, newline=false)
-    lock(print_lock) do
-        timestr = Ryu.writefixed(Float64(elapsedtime/1e9), 6)
-        length(timestr) < 10 && print(" "^(10 - length(timestr)))
-        print(timestr, " seconds")
-        parens = bytes != 0 || allocs != 0 || gctime > 0 || compile_time > 0
-        parens && print(" (")
-        if bytes != 0 || allocs != 0
-            allocs, ma = prettyprint_getunits(allocs, length(_cnt_units), Int64(1000))
-            if ma == 1
-                print(Int(allocs), _cnt_units[ma], allocs==1 ? " allocation: " : " allocations: ")
-            else
-                print(Ryu.writefixed(Float64(allocs), 2), _cnt_units[ma], " allocations: ")
-            end
-            print(format_bytes(bytes))
+    timestr = Ryu.writefixed(Float64(elapsedtime/1e9), 6)
+    str = length(timestr) < 10 ? (" "^(10 - length(timestr))) : ""
+    str *= string(timestr, " seconds")
+    parens = bytes != 0 || allocs != 0 || gctime > 0 || compile_time > 0
+    parens && (str *= " (")
+    if bytes != 0 || allocs != 0
+        allocs, ma = prettyprint_getunits(allocs, length(_cnt_units), Int64(1000))
+        if ma == 1
+            str *= string(Int(allocs), _cnt_units[ma], allocs==1 ? " allocation: " : " allocations: ")
+        else
+            str *= string(Ryu.writefixed(Float64(allocs), 2), _cnt_units[ma], " allocations: ")
         end
-        if gctime > 0
-            if bytes != 0 || allocs != 0
-                print(", ")
-            end
-            print(Ryu.writefixed(Float64(100*gctime/elapsedtime), 2), "% gc time")
-        end
-        if compile_time > 0
-            if bytes != 0 || allocs != 0 || gctime > 0
-                print(", ")
-            end
-            print(Ryu.writefixed(Float64(100*compile_time/elapsedtime), 2), "% compilation time")
-        end
-        parens && print(")")
-        newline && println()
-        nothing
+        str *= format_bytes(bytes)
     end
+    if gctime > 0
+        if bytes != 0 || allocs != 0
+            str *= ", "
+        end
+        str *= string(Ryu.writefixed(Float64(100*gctime/elapsedtime), 2), "% gc time")
+    end
+    if compile_time > 0
+        if bytes != 0 || allocs != 0 || gctime > 0
+            str *= ", "
+        end
+        str *= string(Ryu.writefixed(Float64(100*compile_time/elapsedtime), 2), "% compilation time")
+    end
+    parens && (str *= ")")
+    newline ? println(str) : print(str)
+    nothing
 end
 
 function timev_print(elapsedtime, diff::GC_Diff, compile_time)
     allocs = gc_alloc_count(diff)
-    time_print(elapsedtime, diff.allocd, diff.total_time, allocs, compile_time)
-    print("\nelapsed time (ns): $elapsedtime\n")
+    time_print(elapsedtime, diff.allocd, diff.total_time, allocs, compile_time, true)
+    print("elapsed time (ns): $elapsedtime\n")
     padded_nonzero_print(diff.total_time,   "gc time (ns)")
     padded_nonzero_print(diff.allocd,       "bytes allocated")
     padded_nonzero_print(diff.poolalloc,    "pool allocs")