--- a/base/timing.jl
+++ b/base/timing.jl
@@ -56,12 +56,8 @@ function gc_alloc_count(diff::GC_Diff)
 end
 
 # cumulative total time spent on compilation
-function cumulative_compile_time_ns_before()
-    ccall(:jl_cumulative_compile_time_ns_before, UInt64, ())
-end
-function cumulative_compile_time_ns_after()
-    ccall(:jl_cumulative_compile_time_ns_after, UInt64, ())
-end
+cumulative_compile_time_ns_before() = ccall(:jl_cumulative_compile_time_ns_before, UInt64, ())
+cumulative_compile_time_ns_after() = ccall(:jl_cumulative_compile_time_ns_after, UInt64, ())
 
 # total time spend in garbage collection, in nanoseconds
 gc_time_ns() = ccall(:jl_gc_total_hrtime, UInt64, ())