https://github.com/JuliaLang/julia/pull/38885 intended to only enable compile time during the timing macros, however as @ericphanson [pointed out](https://github.com/JuliaLang/julia/pull/38885#pullrequestreview-553541059) if the code under test throws the timing could be left enabled.

The fix here of using a try-finally has the side effect of enabling the timing report even when errors throw, which didn't happen previously. That could be suppressed if the previous behavior is desired.

master
```
julia> x = rand(1000,1000);

julia> @time x * x;
  0.585421 seconds (2.35 M allocations: 134.031 MiB, 7.19% gc time, 94.24% compilation time)

julia> @time x * x;
  0.015857 seconds (2 allocations: 7.629 MiB)

julia> @time error()
ERROR: 
Stacktrace:
 [1] error()
   @ Base ./error.jl:42
 [2] top-level scope
   @ ./timing.jl:203 [inlined]
 [3] top-level scope
   @ ./REPL[3]:0

julia> # compilation timing is left enabled due to error
```

This PR
```
julia> x = rand(1000,1000);

julia> @time x * x;
  0.677077 seconds (2.36 M allocations: 134.586 MiB, 20.48% gc time, 90.76% compilation time)

julia> @time x * x;
  0.014838 seconds (2 allocations: 7.629 MiB)

julia> @time error()
  0.004712 seconds (406 allocations: 33.688 KiB, 98.85% compilation time)
ERROR: 
Stacktrace:
 [1] error()
   @ Base ./error.jl:42
 [2] top-level scope
   @ ./timing.jl:204 [inlined]
 [3] top-level scope
   @ ./REPL[3]:0
```