@KristofferC 2020-12-18
Are there any race condition concerns when calling `@time` from multiple threads?