@ianshmean 2020-12-18
`@time`'s report printing has never been that well suited to threading

1.5.3
```
julia> x = rand(1000,1000);

julia> Threads.@threads for i in 1:20 @time x * x; end
  0.755160 seconds  0.769467 seconds (2.70 M allocations: 176.013 MiB, 7.66% gc time) (2.72
 M allocations: 176.882 MiB, 7.52% gc time)
  0.782703 seconds (2.73 M allocations: 192.447 MiB, 7.39% gc time)
  0.797099 seconds (2.73 M allocations: 200.083 MiB, 7.26% gc time)
  0.809425 seconds (2.73 M allocations: 207.717 MiB, 7.15% gc time)
  0.822897 seconds (2.73 M allocations: 215.344 MiB, 7.03% gc time)
  0.059053 seconds (  191 allocations: 45.788 MiB0.075888 seconds) (
20.86 k allocations: 39.226 MiB)
  0.084400 seconds (20.87 k allocations: 46.856 MiB)
  0.085214 seconds (20.86 k allocations: 46.855 MiB)
  0.086948 seconds (20.86 k allocations: 46.855 MiB)
  0.086245 seconds (20.85 k allocations: 46.854 MiB)
  0.063872 seconds (150 allocations: 45.785 MiB)
  0.079255 seconds (169 allocations: 38.157 MiB)
  0.085374 seconds (167 allocations: 30.527 MiB)
  0.083243 seconds (167 allocations: 30.527 MiB)
  0.083350 seconds (166 allocations: 22.897 MiB)
  0.083400 seconds (166 allocations: 22.897 MiB)
  0.052417 seconds (101 allocations: 15.264 MiB)
  0.038486 seconds (67 allocations: 7.633 MiB)
```

I'm working on a print lock & compile timer activation lock for this PR