@KristofferC 2020-12-18
I just meant if there is a race condition writing to the variable then (= undefined behaviour).