@ianshmean 2020-12-19
Yes, sorry, I did understand that. I noticed that this PR increases allocations due to the `@time` itself. Looking into it