@ianshmean 2020-12-20
@KristofferC the setting of `jl_measure_compile_time` is now protected by a lock. Thanks for pointing that out.

The report printing is also now protected by a lock to make threaded printing look better.
```
julia> Threads.@threads for i in 1:20 @time x * x; end
  0.013592 seconds (77 allocations: 45.782 MiB, 1509301427.48% compilation time)
  0.030026 seconds (104 allocations: 45.783 MiB, 0.09% compilation time)
  0.062268 seconds (73 allocations: 30.522 MiB, 0.02% compilation time)
  0.045501 seconds (88 allocations: 38.152 MiB, 0.04% compilation time)
  0.078885 seconds (59 allocations: 22.891 MiB, 0.01% compilation time)
  0.094609 seconds (44 allocations: 15.261 MiB)
  0.121956 seconds (217 allocations: 45.788 MiB, 23.85% gc time)
  0.020401 seconds (208 allocations: 45.787 MiB)
  0.028625 seconds (199 allocations: 45.787 MiB)
  0.039365 seconds (189 allocations: 45.786 MiB)
  0.047761 seconds (182 allocations: 45.786 MiB)
  0.057183 seconds (182 allocations: 45.786 MiB)
  0.058982 seconds (174 allocations: 45.785 MiB)
  0.064370 seconds (174 allocations: 45.785 MiB)
  0.067353 seconds (174 allocations: 45.785 MiB)
  0.065737 seconds (173 allocations: 38.156 MiB)
  0.066426 seconds (172 allocations: 30.526 MiB)
  0.068942 seconds (171 allocations: 22.897 MiB)
  0.067828 seconds (170 allocations: 15.267 MiB)
  0.066753 seconds (169 allocations: 7.638 MiB)
```

Although, the first comp time % report here is surprisingly nonsensically high.. Compilation on multiple threads is evidently confusing the singular cumulative counter. I don't see a simple fix. Though I guess it's ok to leave as-is and tackle that in a future PR