@ianshmean 2020-12-30
The compilation timing is now thread-local, without the need for locks, and seems to behave well:
```
julia> x = rand(10,10);

julia> Threads.@threads for i in 1:20 @time x * x; end
  0.626245 seconds (2.35 M allocations: 126.426 MiB, 16.45% gc time, 0.00% compilation time)
  0.626233 seconds (2.35 M allocations: 126.428 MiB, 16.45% gc time, 0.00% compilation time)
  0.626298 seconds (2.35 M allocations: 126.432 MiB, 16.45% gc time, 0.00% compilation time)
  0.626271 seconds (2.35 M allocations: 126.427 MiB, 16.45% gc time, 0.00% compilation time)
  0.626324 seconds (2.35 M allocations: 126.431 MiB, 16.45% gc time, 99.95% compilation time)
  0.626306 seconds (2.35 M allocations: 126.430 MiB, 16.45% gc time, 0.00% compilation time)
  0.000002 seconds (1 allocation: 896 bytes)
  0.000003 seconds (1 allocation: 896 bytes)
  0.000003 seconds (1 allocation: 896 bytes)
  0.000023 seconds (1 allocation: 896 bytes)
  0.000001 seconds (1 allocation: 896 bytes)
  0.000007 seconds (16 allocations: 1.828 KiB)
  0.000004 seconds (10 allocations: 1.312 KiB)
  0.000003 seconds (1 allocation: 896 bytes)
  0.000002 seconds (1 allocation: 896 bytes)
  0.000003 seconds (14 allocations: 1.922 KiB)
  0.000003 seconds (1 allocation: 896 bytes)
  0.000007 seconds (2 allocations: 928 bytes)
  0.000002 seconds (1 allocation: 896 bytes)
  0.000003 seconds (15 allocations: 1.750 KiB)
```
```
julia> x = rand(10,10);

julia> for i in 1:10 @time x * x; end
  0.563264 seconds (2.35 M allocations: 126.422 MiB, 6.58% gc time, 99.94% compilation time)
  0.000006 seconds (1 allocation: 896 bytes)
  0.000006 seconds (1 allocation: 896 bytes)
  0.000002 seconds (1 allocation: 896 bytes)
  0.000002 seconds (1 allocation: 896 bytes)
  0.000002 seconds (1 allocation: 896 bytes)
  0.000002 seconds (1 allocation: 896 bytes)
  0.000001 seconds (1 allocation: 896 bytes)
  0.000001 seconds (1 allocation: 896 bytes)
  0.000002 seconds (1 allocation: 896 bytes)
```